user-status/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
mysqlclient==2.2.7
pika==1.3.2